Move aspect-ratio-dropdown into components
